Restructure directory to match other client libraries.
- API name is `discoveryengine` but the region tag is `genappbuilder` and the product name is Generative AI App Builder.
- Java Repo requested I use the API name for the directories
