+++ .github/blunderbuss.yml
@@ -197,6 +197,7 @@ assign_prs_by:
       - "api: dataproc"
       - "api: dialogflow"
       - "api: enterpriseknowledgegraph"
+      - "api: discoveryengine"
       - "api: language"
       - "api: ml"
       - "api: notebooks"
@@ -298,4 +299,3 @@ assign_prs:
 ###
 # Updates should be made to both assign_issues_by & assign_prs_by sections
 ###
-
