+++ discoveryengine/operations_sample_test.py
@@ -15,9 +15,9 @@
 
 import os
 
-from genappbuilder import get_operation_sample
-from genappbuilder import list_operations_sample
-from genappbuilder import poll_operation_sample
+from discoveryengine import get_operation_sample
+from discoveryengine import list_operations_sample
+from discoveryengine import poll_operation_sample
 
 from google.api_core.exceptions import NotFound
